Restored original version of GPS
I got the video and everything to run correctly on mine machine again, but I am still using the complete path from my machine. I will make it machine independent, but I wanted to get at least get it back to the way it was first. All pages should now be connected in the correct order.